added mid range general added melle soldier

+++ TurnBaseSystems/Assets/Scripts/Units/Abilities/AttackAnimationInfo.cs
@@ -5,5 +5,5 @@ public class AttackAnimationInfo {
     /// <summary>
     /// Time, length of animation
     /// </summary>
-    public int animLength;
+    public float animLength;
 }
